Added Generation URL date-math example
I added an example of using a Generation URL for a csv report that contains date math in the URL, which can cause the Watch to fail if not properly formatted.  This can arise when copying the CSV URL from Kibana but not removing the curly brackets or properly URL-encoding the required characters, and the watch `parse_exception` can be hard to track down.  Users should be aware that the working URL they copy from Kibana may not work in the watch without adjusting things like this.

I wasn't sure if the example expression was too much or too specific to include in a Note format, and am open to re-working if the presentation could be improved.